Found an issue where environment variables would load before they had actually been updated from dotenv, made sure they are always loaded correctly.
