+++ packages/server/src/environment.js
@@ -1,3 +1,17 @@
+const { resolve, join } = require("./utilities/centralPath")
+const { homedir } = require("os")
+const { app } = require("electron")
+
+let LOADED = false
+
+if (!LOADED) {
+  const homeDir = app ? app.getPath("home") : homedir()
+  const budibaseDir = join(homeDir, ".budibase")
+  process.env.BUDIBASE_DIR = budibaseDir
+  require("dotenv").config({ path: resolve(budibaseDir, ".env") })
+  LOADED = true
+}
+
 module.exports = {
   CLIENT_ID: process.env.CLIENT_ID,
   NODE_ENV: process.env.NODE_ENV,
