+++ packages/server/src/utilities/index.js
@@ -1,1 +1,9 @@
 exports.wait = ms => new Promise(resolve => setTimeout(resolve, ms))
+
+exports.isDev = () => {
+  return (
+    process.env.NODE_ENV !== "production" &&
+    process.env.NODE_ENV !== "jest" &&
+    process.env.NODE_ENV !== "cypress"
+  )
+}
